requirements.txt
gradio
google-generativeai